requirements.txt
beautifulsoup4
feedparser
discord.py>=1.0.0
boto3
mutagen
aiohttp>=3.6 # Pythania requires at least 3.6